feat: Store 내 가게 조회 서비스 로직 위치 이동, 기존 StoreCommandServiceImpl -> 변경 StoreQueryServiceImpl
