+++ src/main/java/com/dfdt/delivery/domain/store/application/service/command/StoreCommandService.java
@@ -4,12 +4,10 @@ import com.dfdt.delivery.domain.auth.infrastructure.security.CustomUserDetails;
 import com.dfdt.delivery.domain.store.presentation.dto.request.StoreCreateReqDto;
 import com.dfdt.delivery.domain.store.presentation.dto.request.StoreStatusReqDto;
 import com.dfdt.delivery.domain.store.presentation.dto.request.StoreUpdateReqDto;
-import com.dfdt.delivery.domain.store.presentation.dto.response.MyStoreResDto;
 import com.dfdt.delivery.domain.store.presentation.dto.response.StoreCreateResDto;
 import com.dfdt.delivery.domain.store.presentation.dto.response.StoreStatusResDto;
 import com.dfdt.delivery.domain.store.presentation.dto.response.StoreUpdateResDto;
 
-import java.util.List;
 import java.util.UUID;
 
 public interface StoreCommandService {
@@ -20,8 +18,6 @@ public interface StoreCommandService {
     void deleteStore(UUID storeId, CustomUserDetails user);
 
     void changeIsOpen(UUID storeId, CustomUserDetails userDetails);
-
-    List<MyStoreResDto> getMyStores(String username);
 
     void restoreStore(UUID storeId, CustomUserDetails user);
