+++ src/main/java/com/dfdt/delivery/domain/store/application/service/query/StoreQueryService.java
@@ -1,10 +1,12 @@
 package com.dfdt.delivery.domain.store.application.service.query;
 
+import com.dfdt.delivery.domain.store.presentation.dto.response.MyStoreResDto;
 import com.dfdt.delivery.domain.store.presentation.dto.response.StoreAdminResDto;
 import com.dfdt.delivery.domain.store.presentation.dto.response.StoreResDto;
 import com.dfdt.delivery.domain.store.presentation.dto.response.StoreStatusRequestResDto;
 import org.springframework.data.domain.Page;
 
+import java.util.List;
 import java.util.UUID;
 
 public interface StoreQueryService {
@@ -13,6 +15,8 @@ public interface StoreQueryService {
     Page<StoreResDto> getStores(int page, int size, String sortBy, boolean isAsc, UUID category, String name);
 
     Page<StoreAdminResDto> getStoresAdmin(int page, int size, String sortBy, boolean isAsc, UUID category, String name, Boolean isDeleted);
+
+    List<MyStoreResDto> getMyStores(String username);
 
     Page<StoreStatusRequestResDto> getRequestedStores(int page, int size, String sortBy, boolean isAsc);
 }
